requirements.txt
pyyaml
requests
uritemplate